requirements.txt
RealtimeSTT>=0.3.0
faster-whisper>=1.0.0
openai>=1.0.0
websockets>=12.0
pyaudio>=0.2.14
pyaudiowpatch>=0.2.12
pyyaml>=6.0
numpy>=1.24.0
scipy>=1.11.0
